+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History of versions #
 
+* Version 2.0.0 (2022-12-04)
+  * Added Drag'n'Drop support to move tiles between windows and tabs, or into a new window
+  * Added "Select all" option
+  * Added pointer autohide option
+  * Added reset and clear options
 * Version 1.20.0 (2022-11-20)
   * Fixed Guake window zero height on first run
   * Fixed Guake keybinding failing until re-set